requirements.txt
aggdraw==1.3.12
black==21.12b0
moviepy==1.0.3
pillow==8.4.0
pytest-custom_exit_code==0.3.0
pytest==6.2.5
